Change drawBasicProgressBar parameters from float to int

+++ OPHD/Common.cpp
@@ -395,7 +395,7 @@ NAS2D::StringList split_string(const char *str, char delim)
 }
 
 
-void drawBasicProgressBar(float x, float y, float width, float height, float percent, float padding)
+void drawBasicProgressBar(int x, int y, int width, int height, float percent, int padding)
 {
 	auto& renderer = Utility<Renderer>::get();
 	renderer.drawBox(NAS2D::Rectangle{x, y, width, height}, NAS2D::Color{0, 185, 0});
@@ -403,6 +403,6 @@ void drawBasicProgressBar(float x, float y, float width, float height, float per
 	if (percent > 0.0f)
 	{
 		int bar_width = static_cast<int>(static_cast<float>(width - (padding + padding)) * percent);
-		renderer.drawBoxFilled(NAS2D::Rectangle{x + padding, y + padding + 1.0f, bar_width - 1.0f, height - (padding + padding) - 1.0f}, NAS2D::Color{0, 100, 0});
+		renderer.drawBoxFilled(NAS2D::Rectangle{x + padding, y + padding + 1, bar_width - 1, height - (padding + padding) - 1}, NAS2D::Color{0, 100, 0});
 	}
 }
